my-app/requirements.txt
mkdocs==1.5.3
mkdocs-material==9.5.18
mkdocs-glightbox==0.4.0
mkdocs-git-revision-date-localized-plugin==1.2.8
mkdocs-macros-plugin==1.0.5
feedparser==6.0.11
requests==2.32.3
trafilatura==1.6.3
beautifulsoup4==4.12.3
lxml==5.2.2
Pillow==10.4.0
markdownify==0.14.1
PyYAML==6.0.2
fastapi==0.115.2
uvicorn==0.30.6
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1